Create a AuthProvider component. Create AppProvider component to clean up index.js. Refactor authHeader component to use user context.
