+++ src/utils/misc.js
@@ -1,11 +1,5 @@
 import rottenIcons from 'assets/images/rotten-gas/rottenIcons'
 
-const authHeader = (token) => {
-  return {
-    headers: { Authorization: `bearer ${token}` }
-  }
-}
-
 const rottenReviewImage = (rottenAverage) => (
   rottenAverage > 849 ? rottenIcons.certifiedSmall : 
     rottenAverage > 599 ? rottenIcons.freshSmall :
@@ -15,6 +9,5 @@ const rottenReviewImage = (rottenAverage) => (
 )
 
 export {
-  authHeader,
   rottenReviewImage
 }
